requirements.txt
langchain
tiktoken
python-dotenv
pypdf
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma
langgraph
streamlit
